migrate dataForCharacterSets/dataForCollations from tables to executer
