+++ executor/infoschema_reader.go
@@ -15,6 +15,7 @@ package executor
 
 import (
 	"context"
+	"github.com/pingcap/parser/charset"
 	"sort"
 
 	"github.com/pingcap/parser/model"
@@ -52,6 +53,10 @@ func (e *memtableRetriever) retrieve(ctx context.Context, sctx sessionctx.Contex
 			e.rows = dataForSchemata(sctx, dbs)
 		case infoschema.TableViews:
 			e.rows, err = dataForViews(sctx, dbs)
+		case infoschema.TableCharacterSets:
+			e.rows = dataForCharacterSets()
+		case infoschema.TableCollations:
+			e.rows = dataForCollations()
 		}
 		if err != nil {
 			return nil, err
@@ -152,4 +157,28 @@ func dataForViews(ctx sessionctx.Context, schemas []*model.DBInfo) ([][]types.Da
 		}
 	}
 	return rows, nil
+}
+
+func dataForCharacterSets() (records [][]types.Datum) {
+	charsets := charset.GetSupportedCharsets()
+	for _, charset := range charsets {
+		records = append(records,
+			types.MakeDatums(charset.Name, charset.DefaultCollation, charset.Desc, charset.Maxlen),
+		)
+	}
+	return records
+}
+
+func dataForCollations() (records [][]types.Datum) {
+	collations := charset.GetSupportedCollations()
+	for _, collation := range collations {
+		isDefault := ""
+		if collation.IsDefault {
+			isDefault = "Yes"
+		}
+		records = append(records,
+			types.MakeDatums(collation.Name, collation.CharsetName, collation.ID, isDefault, "Yes", 1),
+		)
+	}
+	return records
 }
